Completed Faults in New implementation - Added current session and MATLAB model rt_OneStep fucn overrun status on CAN - optmised code for size ~70kB - added conditional compilation for Open_FOC and Closed Loop FOC

+++ Core/Inc/NTC_Temp_Reading.h
@@ -3,6 +3,6 @@
 
 #include "main.h"
 
-float NTC_Read(uint16_t analog_val, float R0);
+int16_t NTC_Read(uint16_t analog_val, float R0);
 
 #endif
